+++ lib/formatting.sh
@@ -1,7 +1,19 @@
 __print_colored_content() {
-	[ "${2:0:1}" = "#" ] && BGCOLOR="$2" || BGCOLOR="colour$2"
-	[ "${3:0:1}" = "#" ] && FGCOLOR="$3" || FGCOLOR="colour$3"
-	echo -n "#[fg=${FGCOLOR},bg=${BGCOLOR}]"
-	echo -n "$1"
-	echo -n "#[default]"
+	bgcolor="$2"
+	fgcolor="$3"
+	if [[ $bgcolor == "default" ]] ; then
+		  BGCOLOR=$bgcolor
+	else
+	[ "${bgcolor:0:1}" = "#" ] && BGCOLOR="$bgcolor" || BGCOLOR="colour$bgcolor"
+	fi
+
+	if [[ "$fgcolor" == "default" ]] ; then
+		FGCOLOR=$fgcolor
+	else
+	[ "${fgcolor:0:1}" = "#" ] && FGCOLOR="$fgcolor" || FGCOLOR="colour$fgcolor"
+	fi
+
+	 echo -n "#[fg=${FGCOLOR},bg=${BGCOLOR}]"
+	 echo -n "$1"
+	 echo -n "#[default]"
 }